+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### ✨ Improved
+
+* Do not report a thermistor warning every interval.
+
+
 ## 0.1.2 - 2024-09-18
 
 ### 🚀 New